+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/userdata/ProfileSettingsActivity.kt
@@ -2,27 +2,42 @@ package com.rakuten.tech.mobile.testapp.ui.userdata
 
 import android.app.Activity
 import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
 import android.net.Uri
 import android.os.Bundle
+import android.util.Base64
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.ImageView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.rakuten.tech.mobile.miniapp.testapp.R
+import com.rakuten.tech.mobile.testapp.helper.MiniAppCoroutines
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.android.synthetic.main.profile_settings_activity.*
+import kotlinx.coroutines.Dispatchers.IO
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.io.ByteArrayOutputStream
+import java.lang.Exception
 
 class ProfileSettingsActivity : BaseActivity() {
 
     private lateinit var settings: AppSettings
     private lateinit var profileUrl: String
+    private lateinit var profileUrlBase64: String
+    private val coroutines = MiniAppCoroutines()
+    private val encodePhotoUrlJob: Job = Job()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         settings = AppSettings.instance
         profileUrl = settings.profilePictureUrl
+        profileUrlBase64 = settings.profilePictureUrlBase64
+
         showBackIcon()
         setContentView(R.layout.profile_settings_activity)
         renderProfileSettingsScreen()
@@ -56,6 +71,7 @@ class ProfileSettingsActivity : BaseActivity() {
 
     private fun updateProfile(name: String) {
         settings.profilePictureUrl = profileUrl
+        settings.profilePictureUrlBase64 = profileUrlBase64
         settings.profileName = name.trimEnd()
         finish()
     }
@@ -75,6 +91,30 @@ class ProfileSettingsActivity : BaseActivity() {
             val imageUri = data?.data
             setProfileImage(imageUri)
             profileUrl = imageUri.toString()
+            encodeImageForMiniApp(profileUrl)
+        }
+    }
+
+    private fun encodeImageForMiniApp(profileUrl: String) {
+        coroutines.buildMainScope(encodePhotoUrlJob).launch {
+            try {
+                withContext(IO) {
+                    val uri = Uri.parse(profileUrl)
+                    val inputStream = contentResolver.openInputStream(uri)
+                    inputStream?.use {
+                        val bitmap = BitmapFactory.decodeStream(inputStream)
+                        val byteArrayOutputStream = ByteArrayOutputStream()
+                        bitmap.compress(Bitmap.CompressFormat.JPEG, 50, byteArrayOutputStream)
+                        val bytes: ByteArray = byteArrayOutputStream.toByteArray()
+                        profileUrlBase64 = BASE_64_DATA_PREFIX + Base64.encodeToString(
+                            bytes,
+                            Base64.DEFAULT
+                        )
+                    }
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
         }
     }
 
@@ -85,8 +125,14 @@ class ProfileSettingsActivity : BaseActivity() {
             .into(imageProfile as ImageView)
     }
 
+    override fun onDestroy() {
+        encodePhotoUrlJob.cancel()
+        super.onDestroy()
+    }
+
     companion object {
         private const val PICK_IMAGE = 1001
+        private const val BASE_64_DATA_PREFIX = "data:image/png;base64,"
 
         fun start(activity: Activity) {
             activity.startActivity(Intent(activity, ProfileSettingsActivity::class.java))